break circdep on website-navigation
