+++ src/project/types/website/website-shared.ts
@@ -1,9 +1,8 @@
 /*
-* website-shared.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * website-shared.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { join } from "path/mod.ts";
 
@@ -103,9 +102,9 @@ export function websiteNavigationConfig(project: ProjectContext) {
   let navbar = websiteConfig(kSiteNavbar, project.config) as
     | Navbar
     | undefined;
-  if (typeof (navbar) === "boolean" && navbar) {
+  if (typeof navbar === "boolean" && navbar) {
     navbar = { background: "primary" };
-  } else if (typeof (navbar) !== "object") {
+  } else if (typeof navbar !== "object") {
     navbar = undefined;
   }
 
@@ -114,7 +113,7 @@ export function websiteNavigationConfig(project: ProjectContext) {
   const sidebars =
     (Array.isArray(sidebar)
       ? sidebar
-      : typeof (sidebar) == "object"
+      : typeof sidebar == "object"
       ? [sidebar]
       : undefined) as Sidebar[] | undefined;
 
@@ -213,7 +212,7 @@ export function resolvePageFooter(config?: ProjectConfig) {
   const footerValue = (
     value?: unknown,
   ): string | NavItem[] | undefined => {
-    if (typeof (value) === "string") {
+    if (typeof value === "string") {
       return value as string;
     } else if (Array.isArray(value)) {
       return value as NavItem[];
@@ -224,7 +223,7 @@ export function resolvePageFooter(config?: ProjectConfig) {
 
   const footer: NavigationFooter = {};
   const footerConfig = websiteConfig(kPageFooter, config);
-  if (typeof (footerConfig) === "string") {
+  if (typeof footerConfig === "string") {
     // place the markdown in the center
     footer.center = footerConfig;
   } else if (!Array.isArray(footerConfig)) {
@@ -251,4 +250,98 @@ export function flattenItems(
   };
   sidebarItems.forEach(flatten);
   return items;
+}
+
+export function breadCrumbs(href: string, sidebar?: Sidebar) {
+  if (sidebar?.contents) {
+    const crumbs: SidebarItem[] = [];
+
+    // find the href in the sidebar
+    const makeBreadCrumbs = (href: string, sidebarItems?: SidebarItem[]) => {
+      if (sidebarItems) {
+        for (const item of sidebarItems) {
+          if (item.href === href) {
+            crumbs.push(item);
+            return true;
+          } else {
+            if (item.contents) {
+              if (makeBreadCrumbs(href, item.contents)) {
+                // If this 'section' doesn't have an href, then just use the first
+                // child as the href
+                const breadCrumbItem = { ...item };
+                if (
+                  !breadCrumbItem.href && breadCrumbItem.contents &&
+                  breadCrumbItem.contents.length > 0
+                ) {
+                  breadCrumbItem.href = breadCrumbItem.contents[0].href;
+                }
+
+                crumbs.push(breadCrumbItem);
+                return true;
+              }
+            }
+          }
+        }
+        return false;
+      } else {
+        return false;
+      }
+    };
+    makeBreadCrumbs(href, sidebar.contents);
+    return crumbs.reverse();
+  } else {
+    return [];
+  }
+}
+
+// static navigation (initialized during project preRender)
+export const navigation: Navigation = {
+  sidebars: [],
+};
+
+export function sidebarForHref(href: string, format: Format) {
+  // if there is a single sidebar then it applies to all hrefs
+  if (navigation.sidebars.length === 1) {
+    return navigation.sidebars[0];
+  } else {
+    const explicitSidebar = navigation.sidebars.find((sidebar) => {
+      return sidebar.id === format.metadata[kSiteSidebar];
+    });
+    if (explicitSidebar) {
+      return explicitSidebar;
+    } else {
+      const containingSidebar = navigation.sidebars.find((sidebar) => {
+        return containsHref(href, sidebar.contents);
+      });
+      if (containingSidebar) {
+        return containingSidebar;
+      } else {
+        return undefined;
+      }
+    }
+  }
+}
+
+export function containsHref(href: string, items: SidebarItem[]) {
+  for (let i = 0; i < items.length; i++) {
+    if (items[i].href && items[i].href === href) {
+      return true;
+    } else if (Object.keys(items[i]).includes("contents")) {
+      const subItems = items[i].contents || [];
+      const subItemsHasHref = containsHref(href, subItems);
+      if (subItemsHasHref) {
+        return true;
+      }
+    } else {
+      if (itemHasNavTarget(items[i], href)) {
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+export function itemHasNavTarget(item: SidebarItem, href: string) {
+  return item.href === href ||
+    item.href === href.replace(/\/index\.html/, "/");
 }
